+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog],
 and this project adheres to [Semantic Versioning].
 
+## Unreleased
+
+- added entity id to JEI ingredient when advanced tooltips are enabled
+
 ## [1.1.2] - 2022-10-02
 
 ### Added